project2/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==1.11
django-crispy-forms==1.7.0
django-formtools==2.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
future==0.16.0
httplib2==0.9.2
kombu==4.1.0
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2017.3
redis==2.10.6
six==1.11.0
vine==1.1.4
